backend/orchestrator_agent/pyproject.toml
[project]
name = "orchestrator-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.1.5",
    "langchain-core>=0.1.0",
    "pydantic>=2.0.0",
    "langchain-community>=0.3.19",
    "langchain-qdrant>=0.2.0",
    "langchain-openai>=0.3.9",
    "python-dotenv>=1.0.0",
]

[project.scripts]
orchestrator = "orchestrator:Orchestrator"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
# Configure this as a separate project
managed = true

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.setuptools]
py-modules = ["orchestrator", "tools", "prompt", "main"]
